3-10-2: Changed Fr.h size_a_file() return
I thought twice about data type conversion.  Best to just stick with the implementation-specific data type being used here without having to worry about underflow or overflows.

+++ 3-Internals/Fileroad.h
@@ -80,7 +80,7 @@ char* read_a_file(char* fileName);
 		This function calls stat()
 		The return value is converted from data type off_t	
  */
-long long size_a_file(char* fileName);
+off_t size_a_file(char* fileName);
 
 
 //////////////////////////////////////////////////////////////////////////////
